remove shipped-code from Git repository
This commit removes the contents of the shipped-code folder in bundle
org.eclipse.n4js.external.libraries only from the Git repository. The
shipped code will still be built during the maven build and it will
still be bundled with the IDE product. Also, the contents of top-level
folder n4js-libs are not touched by this commit.